+++ currentVersion.php
@@ -1,5 +1,5 @@
 <?php
 
-	$version = '0.4.0';
+	$version = '0.4.1 (NF)';
 
 ?>